6/12/2021 - 9:01
- Initial support for "input" function to read a line from the stdin file
- Initial support for "remove_from" which remove a symbol from a object
- Initial support for "get_from" which return a value associated with an object
- Added "set" which receives the targeted object, the name of the symbol and the value it will have
- Added "__built_in__" variable which returns an object that has as symbol table the one with all built-in functions
- Added the self symbol table which adds an object with symbol table the one of the vm
- Code reorganization
