+++ pkg/vm/default.go
@@ -5,7 +5,6 @@ import (
 )
 
 /*
-	SetDefaultSymbolTable
 	Type         - (Done)
 	Function     - (Done)
 	Object       - (Done)
@@ -20,21 +19,25 @@ import (
 	Id           - (Done)
 	Range        - (Done)
 	Len          - (Done)
-	Delete		 - ()
+	DeleteFrom   - (Done)
+	Dir          - ()
+	Input        - (Done)
 	ToString     - (Done)
 	ToTuple      - (Done)
 	ToArray      - (Done)
 	ToInteger    - (Done)
 	ToFloat      - (Done)
 	ToBool       - (Done)
-	ToHashTable  - ()
 */
-func (p *Plasma) SetDefaultSymbolTable() {
-	symbolTable := NewSymbolTable(nil)
+func (p *Plasma) setBuiltInSymbols() {
+	/*
+		This is the master symbol table that is protected from writes
+	*/
+	p.builtInSymbolTable = NewSymbolTable(nil)
 
 	// Types
 	type_ := &Type{
-		Object:      p.NewObject(ObjectName, nil, symbolTable),
+		Object:      p.NewObject(ObjectName, nil, p.builtInSymbolTable),
 		Constructor: NewBuiltInConstructor(p.ObjectInitialize),
 		Name:        TypeName,
 	}
@@ -47,12 +50,12 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(TypeName, type_)
+	p.builtInSymbolTable.Set(TypeName, type_)
 	//// Default Error Types
-	exception := p.NewType(RuntimeError, symbolTable, []*Type{type_}, NewBuiltInConstructor(p.RuntimeErrorInitialize))
-	symbolTable.Set(RuntimeError, exception)
-	symbolTable.Set(InvalidTypeError,
-		p.NewType(InvalidTypeError, symbolTable, []*Type{exception},
+	exception := p.NewType(RuntimeError, p.builtInSymbolTable, []*Type{type_}, NewBuiltInConstructor(p.RuntimeErrorInitialize))
+	p.builtInSymbolTable.Set(RuntimeError, exception)
+	p.builtInSymbolTable.Set(InvalidTypeError,
+		p.NewType(InvalidTypeError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -78,8 +81,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(NotImplementedCallableError,
-		p.NewType(NotImplementedCallableError, symbolTable, []*Type{exception},
+	p.builtInSymbolTable.Set(NotImplementedCallableError,
+		p.NewType(NotImplementedCallableError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -112,8 +115,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(ObjectConstructionError,
-		p.NewType(ObjectConstructionError, symbolTable, []*Type{exception},
+	p.builtInSymbolTable.Set(ObjectConstructionError,
+		p.NewType(ObjectConstructionError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -139,8 +142,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(ObjectWithNameNotFoundError,
-		p.NewType(ObjectWithNameNotFoundError, symbolTable, []*Type{exception},
+	p.builtInSymbolTable.Set(ObjectWithNameNotFoundError,
+		p.NewType(ObjectWithNameNotFoundError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -163,8 +166,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 		),
 	)
 
-	symbolTable.Set(InvalidNumberOfArgumentsError,
-		p.NewType(InvalidNumberOfArgumentsError, symbolTable, []*Type{exception},
+	p.builtInSymbolTable.Set(InvalidNumberOfArgumentsError,
+		p.NewType(InvalidNumberOfArgumentsError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -190,8 +193,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(GoRuntimeError,
-		p.NewType(GoRuntimeError, symbolTable, []*Type{exception},
+	p.builtInSymbolTable.Set(GoRuntimeError,
+		p.NewType(GoRuntimeError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -213,8 +216,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(UnhashableTypeError,
-		p.NewType(UnhashableTypeError, symbolTable, []*Type{exception},
+	p.builtInSymbolTable.Set(UnhashableTypeError,
+		p.NewType(UnhashableTypeError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -236,8 +239,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(IndexOutOfRangeError,
-		p.NewType(IndexOutOfRangeError, symbolTable, []*Type{exception},
+	p.builtInSymbolTable.Set(IndexOutOfRangeError,
+		p.NewType(IndexOutOfRangeError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -263,8 +266,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(KeyNotFoundError,
-		p.NewType(KeyNotFoundError, symbolTable, []*Type{exception},
+	p.builtInSymbolTable.Set(KeyNotFoundError,
+		p.NewType(KeyNotFoundError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -297,8 +300,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(IntegerParsingError,
-		p.NewType(IntegerParsingError, symbolTable, []*Type{exception},
+	p.builtInSymbolTable.Set(IntegerParsingError,
+		p.NewType(IntegerParsingError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -316,8 +319,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(FloatParsingError,
-		p.NewType(FloatParsingError, symbolTable, []*Type{exception},
+	p.builtInSymbolTable.Set(FloatParsingError,
+		p.NewType(FloatParsingError, p.builtInSymbolTable, []*Type{exception},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					object.Set(Initialize,
@@ -335,70 +338,197 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
+	p.builtInSymbolTable.Set(BuiltInSymbolProtectionError,
+		p.NewType(BuiltInSymbolProtectionError, p.builtInSymbolTable, []*Type{exception},
+			NewBuiltInConstructor(
+				func(object IObject) *Object {
+					object.Set(Initialize,
+						p.NewFunction(object.SymbolTable(),
+							NewBuiltInClassFunction(object, 1,
+								func(self IObject, arguments ...IObject) (IObject, *Object) {
+									symbolName := arguments[0]
+									if _, ok := symbolName.(*String); !ok {
+										return nil, p.NewInvalidTypeError(symbolName.TypeName(), StringName)
+									}
+									self.SetString(fmt.Sprintf("cannot delete built-in symbol %s", symbolName.GetString()))
+									return p.NewNone(), nil
+								},
+							),
+						),
+					)
+					return nil
+				},
+			),
+		),
+	)
 	//// Default Types
-	symbolTable.Set(NoneName,
-		p.NewType(NoneName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(NoneName,
+		p.NewType(NoneName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(p.NoneInitialize),
 		),
 	)
-	symbolTable.Set(BoolName,
-		p.NewType(BoolName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(BoolName,
+		p.NewType(BoolName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(p.BoolInitialize),
 		),
 	)
-	symbolTable.Set(IteratorName,
-		p.NewType(IteratorName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(IteratorName,
+		p.NewType(IteratorName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(p.IteratorInitialize),
 		),
 	)
-	symbolTable.Set(ObjectName,
-		p.NewType(ObjectName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(ObjectName,
+		p.NewType(ObjectName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(p.ObjectInitialize),
 		),
 	)
-	symbolTable.Set(FunctionName,
-		p.NewType(FunctionName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(FunctionName,
+		p.NewType(FunctionName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(
 				func(object IObject) *Object {
 					return nil
 				}),
 		),
 	)
-	symbolTable.Set(IntegerName,
-		p.NewType(IntegerName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(IntegerName,
+		p.NewType(IntegerName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(p.IntegerInitialize),
 		),
 	)
-	symbolTable.Set(StringName,
-		p.NewType(StringName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(StringName,
+		p.NewType(StringName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(p.StringInitialize),
 		),
 	)
-	symbolTable.Set(BytesName,
-		p.NewType(BytesName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(BytesName,
+		p.NewType(BytesName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(p.BytesInitialize),
 		),
 	)
-	symbolTable.Set(TupleName,
-		p.NewType(TupleName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(TupleName,
+		p.NewType(TupleName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(p.TupleInitialize),
 		),
 	)
-	symbolTable.Set(ArrayName,
-		p.NewType(ArrayName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(ArrayName,
+		p.NewType(ArrayName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(p.ArrayInitialize),
 		),
 	)
-	symbolTable.Set(HashName,
-		p.NewType(HashName, symbolTable, []*Type{type_},
+	p.builtInSymbolTable.Set(HashName,
+		p.NewType(HashName, p.builtInSymbolTable, []*Type{type_},
 			NewBuiltInConstructor(p.HashTableInitialize),
 		),
 	)
 	// Names
 
 	// Functions
-	symbolTable.Set("print",
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set("dir",
+		p.NewFunction(p.builtInSymbolTable,
+			NewBuiltInFunction(1,
+				func(_ IObject, arguments ...IObject) (IObject, *Object) {
+					object := arguments[0]
+					var symbols []IObject
+					for symbol := range object.SymbolTable().Symbols {
+						symbols = append(symbols, p.NewString(p.PeekSymbolTable(), symbol))
+					}
+					return p.NewTuple(p.PeekSymbolTable(), symbols), nil
+				},
+			),
+		),
+	)
+	p.builtInSymbolTable.Set("set",
+		p.NewFunction(p.builtInSymbolTable,
+			NewBuiltInFunction(3,
+				func(_ IObject, arguments ...IObject) (IObject, *Object) {
+					source := arguments[0]
+					symbol := arguments[1]
+					value := arguments[2]
+					if _, ok := symbol.(*String); !ok {
+						return nil, p.NewInvalidTypeError(symbol.TypeName(), StringName)
+					}
+					source.Set(symbol.GetString(), value)
+					return p.NewNone(), nil
+				},
+			),
+		),
+	)
+	p.builtInSymbolTable.Set("get_from",
+		p.NewFunction(p.builtInSymbolTable,
+			NewBuiltInFunction(2,
+				func(_ IObject, arguments ...IObject) (IObject, *Object) {
+					source := arguments[0]
+					symbol := arguments[1]
+					if _, ok := symbol.(*String); !ok {
+						return nil, p.NewInvalidTypeError(symbol.TypeName(), StringName)
+					}
+					value, getError := source.Get(symbol.GetString())
+					if getError != nil {
+						return nil, p.NewObjectWithNameNotFoundError(symbol.GetString())
+					}
+					return value, nil
+				},
+			),
+		),
+	)
+	p.builtInSymbolTable.Set("delete_from",
+		p.NewFunction(p.builtInSymbolTable,
+			NewBuiltInFunction(2,
+				func(_ IObject, arguments ...IObject) (IObject, *Object) {
+					source := arguments[0]
+					symbol := arguments[1]
+					if _, ok := symbol.(*String); !ok {
+						return nil, p.NewInvalidTypeError(symbol.TypeName(), StringName)
+					}
+					_, getError := source.SymbolTable().GetSelf(symbol.GetString())
+					if getError != nil {
+						return nil, p.NewObjectWithNameNotFoundError(symbol.GetString())
+					}
+					delete(source.SymbolTable().Symbols, symbol.GetString())
+					return p.NewNone(), nil
+				},
+			),
+		),
+	)
+	p.builtInSymbolTable.Set("input",
+		p.NewFunction(p.builtInSymbolTable,
+			NewBuiltInFunction(1,
+				func(_ IObject, arguments ...IObject) (IObject, *Object) {
+					message := arguments[0]
+					var messageString IObject
+					if _, ok := message.(*String); !ok {
+						messageToString, getError := message.Get(ToString)
+						if getError != nil {
+							return nil, p.NewObjectWithNameNotFoundError(ToString)
+						}
+						if _, ok = messageToString.(*Function); !ok {
+							return nil, p.NewInvalidTypeError(messageToString.TypeName(), FunctionName)
+						}
+						toStringResult, callError := p.CallFunction(messageToString.(*Function), p.PeekSymbolTable())
+						if callError != nil {
+							return nil, callError
+						}
+						if _, ok = toStringResult.(*String); !ok {
+							return nil, p.NewInvalidTypeError(toStringResult.TypeName(), StringName)
+						}
+						messageString = toStringResult
+					} else {
+						messageString = message
+					}
+					_, writingError := p.StdOut().Write([]byte(messageString.GetString()))
+					if writingError != nil {
+						return nil, p.NewGoRuntimeError(writingError)
+					}
+					if p.StdInScanner().Scan() {
+						return p.NewString(p.PeekSymbolTable(), p.StdInScanner().Text()), nil
+					}
+					return nil, p.NewGoRuntimeError(p.StdInScanner().Err())
+				},
+			),
+		),
+	)
+	p.builtInSymbolTable.Set("print",
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					value := arguments[0]
@@ -422,8 +552,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set("println",
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set("println",
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					value := arguments[0]
@@ -447,19 +577,18 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set("id",
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set("id",
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					object := arguments[0]
-					// ToDo: Fix me, what about those times the id is greater than int64 max value
-					return p.NewInteger(p.PeekSymbolTable(), int64(object.Id())), nil
+					return p.NewInteger(p.PeekSymbolTable(), object.Id()), nil
 				},
 			),
 		),
 	)
-	symbolTable.Set("hash",
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set("hash",
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					object := arguments[0]
@@ -475,8 +604,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set("range",
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set("range",
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(3,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					start := arguments[0]
@@ -527,8 +656,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set("len",
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set("len",
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					object := arguments[0]
@@ -552,8 +681,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 		),
 	)
 	// To... (Transformations)
-	symbolTable.Set(ToFloat,
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set(ToFloat,
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					toFloat, getError := arguments[0].Get(ToFloat)
@@ -568,8 +697,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(ToString,
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set(ToString,
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					toString, getError := arguments[0].Get(ToString)
@@ -584,8 +713,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(ToInteger,
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set(ToInteger,
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					toInteger, getError := arguments[0].Get(ToInteger)
@@ -600,8 +729,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(ToArray,
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set(ToArray,
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					// First check if it is iterable
@@ -618,8 +747,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(ToTuple,
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set(ToTuple,
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					// First check if it is iterable
@@ -636,8 +765,8 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	symbolTable.Set(ToBool,
-		p.NewFunction(symbolTable,
+	p.builtInSymbolTable.Set(ToBool,
+		p.NewFunction(p.builtInSymbolTable,
 			NewBuiltInFunction(1,
 				func(_ IObject, arguments ...IObject) (IObject, *Object) {
 					toBool, getError := arguments[0].Get(ToBool)
@@ -652,5 +781,5 @@ func (p *Plasma) SetDefaultSymbolTable() {
 			),
 		),
 	)
-	p.programMasterSymbolTable = symbolTable
+	p.builtInSymbolTable = p.builtInSymbolTable
 }
